code-structure/pyproject.toml
[project]
name = "code-structure"
version = "0.1.0"
description = "AI Project for generating technical structure of the ABAP code"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-ollama>=0.3.6",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "transformers>=4.54.1",
]

[dependency-groups]
dev = ["ruff>=0.12.7"]
